Remove requirement that script be located in $PROJECT_DIR
This is accomplished by avoiding `cd` as much as possible (git -C is helpful
here), and by storing the script's current working directory initially, so
that if the path does get changed it won't cause problems.

A few places where the existing config variables for the core location weren't
being used were cleaned up as well.